+++ pkg/record_use/lib/src/internal/definition.dart
@@ -21,15 +21,12 @@ class Definition {
   factory Definition.fromJson(
     Map<String, dynamic> json,
     List<Identifier> identifiers,
+    List<String> uris,
   ) {
     final identifier = identifiers[json['id'] as int];
     return Definition(
       identifier: identifier,
-      location: Location.fromJson(
-        json['@'] as Map<String, dynamic>,
-        identifier.uri,
-        null,
-      ),
+      location: Location.fromJson(json['@'] as Map<String, dynamic>, uris),
       loadingUnit: json['loadingUnit'] as String?,
     );
   }
@@ -40,7 +37,7 @@ class Definition {
   ) =>
       {
         'id': identifiers[identifier]!,
-        '@': location.toJson(),
+        '@': location.toJson(uris),
         'loadingUnit': loadingUnit,
       };