+++ pkg/record_use/test_data/drop_dylib_recording/lib/src/drop_dylib_recording.dart
@@ -6,19 +6,23 @@ import 'package:meta/meta.dart';
 
 import 'drop_dylib_recording_bindings.dart' as bindings;
 
-@RecordUse()
-void getMathMethod(String symbol) {
-  if (symbol == 'add') {
-    print('Hello world: ${_MyMath.add(3, 4)}!');
-  } else if (symbol == 'multiply') {
-    print('Hello world: ${_MyMath.multiply(3, 4)}!');
-  } else {
-    throw ArgumentError('Must pass either "add" or "multiply"');
-  }
-}
-
-class _MyMath {
+class MyMath {
+  @RecordUse()
   static int add(int a, int b) => bindings.add(a, b);
 
+  @RecordUse()
   static int multiply(int a, int b) => bindings.multiply(a, b);
+
+  @RecordCallToC('add')
+  static int double(int a) => bindings.add(a, a);
+
+  @RecordCallToC('multiply')
+  static int square(int a) => bindings.multiply(a, a);
+}
+
+@RecordUse()
+class RecordCallToC {
+  final String symbol;
+
+  const RecordCallToC(this.symbol);
 }