+++ pkg/vm/testcases/transformations/record_use/partfile_main.dart
@@ -2,8 +2,9 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:drop_dylib_recording/drop_dylib_recording.dart';
+import 'package:meta/meta.dart' show RecordUse;
+part 'partfile_helper.dart';
 
-void main(List<String> arguments) {
-  getMathMethod('add');
+void main() {
+  print(SomeClass.someStaticMethod(42));
 }