+++ pkg/record_use/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.3.0
+
+- Make `InstanceConstant` a `Constant`.
+- Separate import from location uri.
+
 ## 0.2.0
 
 - Use maps instead of lists in serialization.